Fixed problems with V2 not handling UTF-8 characters by doing conversion from UTF-8 to Latin-1 in a new V2Localisation class. The bug can be observed with a CK2-1066 converted game where the accented o of León is not rendered at all.

+++ EU4toV2/Source/V2World/V2Flags.cpp
@@ -83,7 +83,7 @@ void V2Flags::SetV2Tags(const std::map<std::string, V2Country*>& V2Countries)
 	}
 
 	// All the remaining tags now need one of the usable flags.
-	static std::mt19937 generator(std::chrono::system_clock::now().time_since_epoch().count());
+	static std::mt19937 generator(static_cast<int>(std::chrono::system_clock::now().time_since_epoch().count()));
 	size_t mappingsMade = 0;
 	for (std::set<std::string>::const_iterator i = requiredTags.cbegin(); i != requiredTags.cend(); ++i)
 	{
